# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API proyecto EVA",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/planilla_seguimiento": {
            "get": {
                "tags": [
                    "planillas Seguimientos"
                ],
                "summary": "Obtiene una lista de los planillas seguimientos",
                "operationId": "4550bda1cc291d34ce63064f93079066",
                "responses": {
                    "200": {
                        "description": "Lista de planillas seguimientos"
                    }
                }
            },
            "post": {
                "tags": [
                    "planillas Seguimientos"
                ],
                "summary": "Crear un nuevo planilla Seguimiento",
                "operationId": "7a89b3fd58954de6d679d4d7a169dd38",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "fecha_revision",
                                    "hora_revision",
                                    "id_empresa"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Item 1"
                                    },
                                    "fecha_revision": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024/8/10)"
                                    },
                                    "hora_revision": {
                                        "type": "time",
                                        "example": "18:00"
                                    },
                                    "id_empresa": {
                                        "type": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "planilla Seguimiento creado con éxito"
                    },
                    "422": {
                        "description": "Problemas con los datos ingresados"
                    }
                }
            }
        },
        "/api/planilla_seguimiento/{id}": {
            "get": {
                "tags": [
                    "planillas Seguimientos"
                ],
                "summary": "Mostar un planilla Seguimiento",
                "operationId": "5dab0a2d77b7a18f873e5fd6b833a4b5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del planilla Seguimiento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del planilla Seguimiento"
                    },
                    "404": {
                        "description": "planilla Seguimiento no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "planillas Seguimientos"
                ],
                "summary": "Actualizar un planilla Seguimiento",
                "operationId": "a65d68397f8a3641abbf3738f890361d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del planilla Seguimiento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Item 1"
                                    },
                                    "fecha_revision": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024/8/10)"
                                    },
                                    "hora_revision": {
                                        "type": "time",
                                        "example": "18:00"
                                    },
                                    "id_empresa": {
                                        "type": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "planilla Seguimiento actualizado con éxito"
                    },
                    "404": {
                        "description": "planilla Seguimiento no encontrado"
                    },
                    "422": {
                        "description": "Problemas con los datos ingresados"
                    }
                }
            },
            "delete": {
                "tags": [
                    "planillas Seguimientos"
                ],
                "summary": "Eliminar un planilla Seguimiento",
                "operationId": "87bb1f381bdaff418eae2ea131a99dc3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del planilla Seguimiento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "planilla Seguimiento eliminado"
                    },
                    "404": {
                        "description": "planilla Seguimiento no encontrado"
                    }
                }
            }
        },
        "/api/planilla_seguimiento/empresa/{id}": {
            "get": {
                "tags": [
                    "planillas Seguimientos"
                ],
                "summary": "Mostar un planilla Seguimiento",
                "operationId": "16be7e8b9b1b461db43ba3d217cb0e3c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del empresa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del planilla Seguimiento por empresa"
                    },
                    "404": {
                        "description": "Empresa no encontrada"
                    }
                }
            }
        },
        "/api/planilla_seguimiento/semanal/{idUsuario}": {
            "get": {
                "tags": [
                    "planillas Seguimientos"
                ],
                "summary": "Mostar planillas Seguimientos por docente",
                "operationId": "bfe745f507e29ceee65e862596bc018e",
                "parameters": [
                    {
                        "name": "idUsuario",
                        "in": "path",
                        "description": "ID de el docente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del planilla Seguimiento por docente"
                    },
                    "404": {
                        "description": "docente no encontrada"
                    }
                }
            }
        },
        "/api/planilla_seguimiento/asistencia/": {
            "get": {
                "tags": [
                    "planillas Seguimientos"
                ],
                "summary": "Mostar planillas Seguimientos con asistencia",
                "operationId": "23e058374ad6ad8d8c669b99567612db",
                "parameters": [
                    {
                        "name": "idUsuario",
                        "in": "path",
                        "description": "ID de la plantilla",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del planilla Seguimiento con asistencia"
                    },
                    "404": {
                        "description": "planilla no encontrada"
                    }
                }
            }
        },
...
    },
...
}